Updated Inventory to handle double

+++ Server/Inventory/InventoryItem.cs
@@ -48,7 +48,7 @@ namespace Server.Inventory
             }
         }
 
-        public InventoryItem(string id, string customName, string itemValue = null, int quantity = 1)
+        public InventoryItem(string id, string customName, string itemValue = null, double quantity = 1)
         {
             Id = id;
             CustomName = customName;
@@ -56,7 +56,7 @@ namespace Server.Inventory
             Quantity = quantity;
         }
 
-        public float GetTotalWeight(int quantity = 1) => (float)Math.Round(ItemInfo.Weight * quantity, 2);
+        public float GetTotalWeight(double quantity = 1) => (float)Math.Round(ItemInfo.Weight * quantity, 2);
 
         //public float GetTotalCapacity
         //{
